Alpert rule
The rules defined here assume that the integral is to be taken over the interval [0,1]. The interval is divided into N+1 intervals. The leftmost and rightmost intervals are handled in a special way, depending on whether a particular kind of singularity is expected.

A singularity may exist at the left endpoint, x = 0. The cases are:

regular, no singularity;
power, the integrand has the form g(x)=x^(-1/2)*phi(x)+psi(x);
log, the integrand has the form g(x)=phi(x)*log(x)+psi(x);
In case one, the regular Alpert rule is used in both end intervals. In case two, the power singular Alpert rule is used in the leftmost interval. In case three, the log singular Alpert rule is used in the leftmost interval.