Alpert rule
alpert_rule, a Python code which has tabulated values that define Alpert quadrature rules of a number of orders of accuracy for functions that are regular, log singular, or power singular.

The rules defined here assume that the integral is to be taken over the interval [0,1]. The interval is divided into N+1 intervals. The leftmost and rightmost intervals are handled in a special way, depending on whether a particular kind of singularity is expected.

A singularity may exist at the left endpoint, x = 0. The cases are:
1. regular, no singularity;
2. power, the integrand has the form g(x)=x^(-1/2)*phi(x)+psi(x);
3. log, the integrand has the form g(x)=phi(x)*log(x)+psi(x);

In case one, the regular Alpert rule is used in both end intervals. In case two, the power singular Alpert rule is used in the leftmost interval. In case three, the log singular Alpert rule is used in the leftmost interval.

Reference:
Bradley Alpert,
Hybrid Gauss-Trapezoidal Quadrature Rules,
SIAM Journal on Scientific Computing,
Volume 20, Number 5, pages 1551-1584, 1999.