Analemma
analemma, a Python code which evaluates the equation of time.

The code computes and plots the equation of time, the declination, and an analemma curve for various orbital parameters. The analemma is the curve traced by the position of the sun, measured at clock noon, over a year.

Usage:
analemma ( ecc, lon, obliq )
where the optional input parameters are:
1. ecc, the eccentricity;
2. lon, the perihelion longitude in degrees;
3. obliq, the axial obliquity in degrees;

The code with no options uses the following default values:

1. eccentricity = 0.01671;
2. longitude of perihelion = 1.347 radians = 77.18 degrees;
3. axial obliquity = 0.4091 radians = 23.44 degrees;

Reference:
Brian Tung,
Figure Eight in the Sky, a new perspective on an old fascination,
Astronomical Games, August 2002,
https://www.astronomycorner.net/games/analemma.html.