asa006
asa006, a Python code which computes the Cholesky factor of a symmetric positive definite (SPD) matrix.

This is a version of Applied Statistics Algorithm 6.

If A is a symmetric positive definite matrix, then there is an upper triangular matrix U with the property that

A = U' * U
The matrix U is known as the Cholesky factor of A, and can be used to easily solve linear systems involving A or compute the inverse of A.
The algorithm implemented here uses a compressed storage for both the matrix A and the factor U. This saves some storage, but can make computations a little awkward.

Reference:
1. PR Freeman,
Remark AS R44: A Remark on AS 6 and AS7: Triangular decomposition of a symmetric matrix and Inversion of a positive semi-definite symmetric matrix,
Applied Statistics,
Volume 31, Number 3, 1982, pages 336-339.
2. Michael Healy,
Algorithm AS 6: Triangular decomposition of a symmetric matrix,
Applied Statistics,
Volume 17, Number 2, 1968, pages 195-197.