asa007
asa007, a Python code which computes the inverse of a symmetric positive definite (SPD) matrix, by Michael Healy.

This is a version of Applied Statistics Algorithm 7.

The algorithm implemented here uses a compressed storage for both the matrix A and the factor U. This saves some storage, but can make computations a little awkward.

Reference:
Michael Healy,
Algorithm AS 7: Inversion of a Positive Semi-Definite Symmetric Matrix,
Applied Statistics,
Volume 17, Number 2, 1968, pages 198-199.