asa109
asa109, a Python code which inverts the incomplete Beta function by Cran, Martin and Thomas. This is a version of Applied Statistics Algorithm 109.

Given a value of the incomplete Beta function, it can determine the argument that produced that value.

Reference:
GW Cran, KJ Martin, GE Thomas,
Remark AS R19 and Algorithm AS 109: A Remark on Algorithms AS 63: The Incomplete Beta Integral and AS 64: Inverse of the Incomplete Beta Integeral,
Applied Statistics,
Volume 26, Number 1, 1977, pages 111-114.