asa245
asa245, a Python code which evaluates the logarithm of the gamma function, by Allan Mcleod;

This is a version of Applied Statistics Algorithm 245.

Reference:
1. Cornelius Lanczos,
A precision approximation of the gamma function,
SIAM Journal on Numerical Analysis, B,
Volume 1, 1964, pages 86-96.
2. Allan Mcleod,
Algorithm AS 245: A Robust and Reliable Algorithm for the Logarithm of the Gamma Function,
Applied Statistics,
Volume 38, Number 2, 1989, pages 397-402.