atbash
atbash, a Python code which applies the Atbash substitution cipher to a string of text.

The Atbash cipher replaces each character by its reverse element in the alphabet. In particular 'A' becomes 'Z', 'B' becomes 'Y' and so on. Case is respected. Nonalphabetic characters are unchanged. The Atbash cipher is symmetric, so applying it to an enciphered string returns the original string.