ball_grid
ball_grid, a Python code which generates a grid of points over the interior of an arbitrary ball in 3D.

The code allows the user to define an arbitrary ball by choosing its radius and center. The user also choose N, the number of subintervals into which a horizontal radius line will be subdivided. It then returns the gridpoints defined by that choice.