ball_integrals
ball_integrals, a Python code which returns the exact value of the integral of any monomial over the interior of the unit ball in 3D.

The interior of the unit ball in 3D is defined by

        x^2 + y^2 + z^2 <= 1
      
The integrands are all of the form

        f(x,y,z) = x^e1 * y^e2 * z^e3
      
where the exponents are nonnegative integers. If any exponent is an odd integer, the integral will be zero. Thus, the "interesting" results occur when all exponents are even.