barycentric_interp_1d
barycentric_interp_1d, a Python code which defines and evaluates the Lagrange polynomial p(x) which interpolates a set of data, so that p(x(i)) = y(i). Because a barycentric formulation is used, polynomials of very high degree can safely be used.

Efficient calculation of the barycentric polynomial interpolant requires that the function to be interpolated be sampled at points from a known family, for which the interpolation weights have been precomputed. Such families include

evenly spaced points (but this results in an ill-conditioned system);
Chebyshev Type 1 points;
Chebyshev Type 2 points;
Chebyshev Type 3 points;
Chebyshev Type 4 points;
and any linear mapping of these points to an arbitary interval [A,B].
Note that in the Berrut/Trefethen reference, there is a significant typographical error on page 510, where an adjustment is made in cases where the polynomial is to be evaluated exactly at a data point. The paper reads:

          exact(xdiff==0) = 1;
        
but it should read
          exact(xdiff==0) = j;