bisection_rc
bisection_rc, a Python code which demonstrates the simple bisection method for solving a scalar nonlinear equation in a change of sign interval, using reverse communication (RC).

The code assumes that an interval [a,b] is known, over which the function f(x) is continuous, and for which f(a) and f(b) are of opposite sign. By repeatedly computing and testing the midpoint, the halving change of sign interval may be reduced, so that either the uncertainty interval or the magnitude of the function value becomes small enough to satisfy the user as an approximation to the location of a root of the function.

This routine is in part a demonstration of the idea of reverse communication. Many zero finders require that the user define f(x) by writing a function with a very specific set of input and output arguments, and sometimes with a specific name, so that the user can call the zero finder, which in turn can call the function. This is sometimes an awkward formulation to follow. Reverse communication instead allows the user to retain control of the evaluation.

To use the reverse communication zero finder, the user defines the values of A and B, and sets a parameter JOB to zero to indicate that this is the first call. From then on, the zero finder repeatedly returns a value X, asking the user to evaluate the function there. Once the user has evaluated FX = f(X), the user may accept this approximation to the root, or else call the zero finder again, passing the just-computed value of FX so that it can take another bisection step.