boundary_locus
boundary_locus, a Python code which uses the boundary locus method to display the region of absolute stability for an ODE solution method, based on a procedure by Randall Leveque.

The analysis begins by applying the ODE solution method to the model equation:

y' = lambda * y
then letting z represent the product h*lambda, where h is the stepsize, and reformulating the solution method as
Ynew = R(z) * Yold
The method will be absolutely stable if |R(z)| < 1. We can then simply create a contour plot of the function |R(z)| to display the stability region.