c8lib
c8lib, a Python code which implements certain elementary functions for "C8" or double precision complex variables.

The symbol "C8" is used to suggest a double precision complex variable. Here the 8 indicates that the underlying pair of real numbers each use 8 bytes of storage.

C8LIB contains routines that operate on scalars, vectors or doubly dimensioned arrays of C8's. A vector of C8's is denoted a C8VEC, and a doubly dimensioned array is a C8MAT.