calpak
calpak, a Python code which computes various simple calendrical quantities.

These include converting from day-number/year to day/month/year format, calculating the time difference between two dates, finding the day of the week of a given date, and other feats.

Some common methods of marking the date include:

JED, the Julian Ephemeris Date, a count of days since a long time ago;
YMD, for "year, month, day";
YJ, for "year, day number".
While there have been many calendars over the years, it is instructive to contemplate just the crazy story of our current "common" calendar. To this day, people disagree about whether there was a year 0, although the Julian calendar was a Roman invention, and Dionysius Exiguus, who gets the blame for shifting the Julian calendar's starting date to the birth year of Christ four hundred years afterwards, didn't have an accurate idea of when that was.

There was a controversial shift from the Julian to the Gregorian calendar, which took place piecemeal throughout the Catholic world, with several countries actually switching back and forth more than once, and with England holding out on the old system until after George Washington was born (which means he was born on February 11 AND February 22).

People didn't agree on when the year started, so that January and February, in particular, were a little murky about which year they belonged to, and the year sometimes started around March 22, near the vernal equinox.

There are a number of side issues, including

determining the day of the week of a given date;
determining the number of days between two dates;
the computation of the date of certain holidays;