chebyshev
chebyshev, a Python code which constructs the Chebyshev interpolant to a function.

Note that the user is not free to choose the interpolation points. Instead, the function f(x) will be evaluated at points chosen by the algorithm. In the standard case, in which the interpolation interval is [-1,+1], these points will be the zeros of the Chebyshev polynomial of order N. However, the algorithm can also be applied to an interval of the form [a,b], in which case the evaluation points are linearly mapped from [-1,+1].

The resulting interpolant is defined by a set of N coefficients c(), and has the form:

        C(f)(x) = sum ( 1 <= i <= n ) c(i) T(i-1,x) - 0.5 * c(1)
      
where T(i-1,x) is the (i-1)-th Chebyshev polynomial.
Within the interval [-1,+1], or the generalized interval [a,b], the interpolant actually remains bounded by the sum of the absolute values of the coefficients c(). It is therefore common to use Chebyshev interpolants as approximating functions over a given interval.