circle_integrals
circle_integrals, a Python code which returns the exact value of the integral of any monomial along the circumference of the unit circle in 2D.

The circumference of the unit circle in 2D is defined by

        x^2 + y^2 = 1
      
The integrands are all of the form

        f(x,y) = x^e1 * y^e2
      
where the exponents are nonnegative integers. If any exponent is an odd integer, the integral will be zero. Thus, the "interesting" results occur when all exponents are even.