condition
condition, a Python code which implements methods for computing or estimating the condition number of a matrix.

Let ||*|| be a matrix norm, let A be an invertible matrix, and inv(A) the inverse of A. The condition number of A with respect to the norm ||*|| is defined to be

        kappa(A) = ||A|| * ||inv(A)||
      
If A is not invertible, the condition number is taken to be infinity.

Facts about the condition number include:

1 <= kappa(A) for all matrices A.
1 = kappa(I), where I is the identity matrix.
for the L2 matrix norm, the condition number of any orthogonal matrix is 1.
for the L2 matrix norm, the condition number is the ratio of the maximum to minimum singular values;