cube_grid
cube_grid, a Python code which generates a grid of points over the interior of a cube (actually, a hexahedron) in 3D.

The grid is created inside a 3D shape which we may informally call a cube. Actually, it is a hexahedron (six-sided shape) determined by 8 user-specified vertices, given in a particular order that is suggested by the following list:

         000
         X00
         0Y0
         XY0
         00Z
         X0Z
         0YZ
         XYZ
       
For each dimension of the three spatial dimensions, the user may specify the number of subintervals into which the edges of the hexahedron are to be divided (at least 1, in each case). The total number of grid points may be determined by adding one to each of these three numbers and computing their product.