cuda_loop
cuda_loop, a Python code which demonstrates how the user's choice of CUDA blocks and threads determines how the user's tasks will be distributed across the GPU.

A CUDA kernel "kernel()" is invoked by a command of the form

    
      kernel << blocks, threads >> ( args )
      
where blocks and threads are each vectors of up to 3 values, listing the number of blocks and number of threads to be used.
If a problem involves N tasks, then tasks are allotted to specific CUDA processes in an organized fashion. Some processes may get no tasks, one task, or multiple tasks.

Each process is given variables that can be used to determine the tasks to be performed:

gridDim.x, gridDim.y, gridDim.z: the block dimensions as given by the user in "blocks";
blockDim.x, blockDim.y, blockDim.z: the thread dimensions as given by the user in "threads";
blockIdx.x, blockIdx.y, blockId.z: the block indices for this process.
threadIdx.x, threadIdx.y, threadIdx.z: the thread indices for this process.
Essentially, a process can determine its linear index K by:

      K = threadIdx.x
        +  blockdim.x  * threadIdx.y
        +  blockDim.x  *  blockDim.y  * threadIdx.z
        +  blockDim.x  *  blockDim.y  *  blockDim.z  * blockIdx.x
        +  blockDim.x  *  blockDim.y  *  blockDim.z  *  gridDim.x  * blockIdx.y
        +  blockDim.x  *  blockDim.y  *  blockDim.z  *  gridDim.x  *  gridDim.y  * blockIdx.z
      
It should use this index as follow:
      Set task T = K.

      while ( T < N )
        carry out task T;
        T = T + blockDim.x * blockDim.y * blockDim.z * gridDim.x * gridDim.y * gridDim.z.
      
The CUDA_LOOP program suggests how a specific set of block and thread parameters would determine the assignment of individual tasks to CUDA processes.