disk_grid
disk_grid, a Python code which generates a grid of points over the interior of a disk in 2D.

The code allows the user to define an arbitrary disk by choosing its radius and center. The user also chooses N, the number of subintervals into which the horizontal radius line will be subdivided. It then returns the gridpoints defined by that choice.

Another function allows the user to specify that N points on a Fibonacci spiral are to be placed within the disk.

The test program creates input files to GNUPLOT which will create graphic images of the grids.