disk_integrals
disk_integrals, a Python code which returns the exact value of the integral of any monomial over the interior of a disk of radius R centered at the origin in 2D.

The interior of the disk is defined by

        x^2 + y^2 <= r^2
      
The integrands are all of the form

        f(x,y) = x^e1 * y^e2
      
where the exponents are nonnegative integers. If any exponent is an odd integer, the integral will be zero. Thus, the "interesting" results occur when all exponents are even.