ellipse_grid
ellipse_grid, a Python code which generates a grid of points over the interior of an ellipse in 2D.

The code allows the user to define an ellipse by choosing its center and the length of its major and minor axes. The user also choose N, the number of subintervals into which the minor radius will be subdivided. It then returns the gridpoints defined by that choice.