elliptic_integral
elliptic_integral, a Python code which evaluates elliptic integral functions using Carlson's elliptic functions.

The complete and incomplete elliptic integrals of the first, second and third kind can be evaluated, with parameters A (angle in degrees), K (sine of A) or M (the modulus, K^2).

Routines are also supplied to evaluate Jacobi's elliptic functions CN, DN and SN.