fd1d_advection_lax_wendroff
fd1d_advection_lax_wendroff, a Python code which applies the finite difference method (FDM) to solve the time-dependent advection equation ut = - c * ux in one spatial dimension, with a constant velocity, using the Lax-Wendroff method for the time derivative, creating a graphics file using matplotlib.

The Lax-Wendroff method is a modification to the Lax method with improved accuracy.

We solve the constant-velocity advection equation in 1D,

        du/dt = - c du/dx
      
over the interval:
        0.0 <= x <= 1.0
      
with periodic boundary conditions, and with a given initial condition
        u(0,x) = (10x-4)^2 (6-10x)^2 for 0.4 <= x <= 0.6
               = 0 elsewhere.
      
For our simple case, the advection velocity is constant in time and space. Therefore, (given our periodic boundary conditions), the solution should simply move smoothly from left to right, returning on the left again. While the Lax method produces an artificial smearing of the solution because of an artificial viscosity effect, this behavior is much reduced for the Lax-Wendroff method.