fd1d_bvp
fd1d_bvp, a Python code which applies the finite difference method (FDM) to solve a two point boundary value problem (BVP) in one spatial dimension.

The boundary value problem (BVP) that is to be solved has the form:

        - d/dx ( a(x) * du/dx ) + c(x) * u(x) = f(x)
      
in the interval X(1) < x < X(N). The functions a(x), c(x), and f(x) are given functions, and a formula for a'(x) is also available.
Boundary conditions are applied at the endpoints, and in this case, these are assumed to have the form:

        u(X(1)) = 0.0;
        u(X(N)) = 0.0.
      
To compute a finite difference approximation, a set of n nodes is defined over the interval, and, at each interior node, a discretized version of the BVP is written, with u''(x) and u'(x) approximated by central differences.