fd1d_heat_explicit
fd1d_heat_explicit, a Python code which solves the time-dependent 1D heat equation, using the finite difference method in space, and an explicit version of the method of lines to handle integration in time.

This code solves

        dUdT - k * d2UdX2 = F(X,T)
      
over the interval [A,B] with boundary conditions
        U(A,T) = UA(T),
        U(B,T) = UB(T),
      
over the time interval [T0,T1] with initial conditions
        U(X,T0) = U0(X)
      
A second order finite difference is used to approximate the second derivative in space.

The solver applies an explicit forward Euler approximation to the first derivative in time.

The resulting finite difference form can be written as

       U(X,T+dt) - U(X,T)         ( U(X-dx,T) - 2 U(X,T) + U(X+dx,T) )
       ------------------  = k *  ------------------------------------ + F(X,T)
                dt                          dx * dx
      
or, assuming we have solved for all values of U at time T, we have
       U(X,T+dt) = U(X,T) + cfl * ( U(X-dx,T) - 2 U(X,T) + U(X+dx,T) ) + dt * F(X,T)
      
where "cfl" is the Courant-Friedrichs-Loewy coefficient:
        cfl = k * dt / dx / dx
      
In order for accurate results to be computed by this explicit method, the cfl coefficient must be less than 0.5!
Other approaches would involve a fully implicit backward Euler approximation or the Crank-Nicholson approximation. These latter two methods have improved stability.

A second worthwhile change would be to replace the constant heat conductivity K by a function K(X,T). The spatial variation would allow for the modeling of a region divided into subregions of different materials.