fem1d_bvp_linear
fem1d_bvp_linear, a Python code which applies the finite element method, with piecewise linear elements, to a two point boundary value problem in one spatial dimension, and compares the computed and exact solutions with the L2 and seminorm errors.

The boundary value problem (BVP) that is to be solved has the form:

        - d/dx ( a(x) * du/dx ) + c(x) * u(x) = f(x)
      
in the interval 0 < x < 1. The functions a(x), c(x), and f(x) are given.
Boundary conditions are applied at the endpoints, and in this case, these are assumed to have the form:

        u(0.0) = 0.0;
        u(1.0) = 0.0.
      
To compute a finite element approximation, a set of n equally spaced nodes is defined from 0.0 to 1.0, a set of piecewise linear basis functions is set up, with one basis function associated with each node, and then an integral form of the BVP is used, in which the differential equation is multiplied by each basis function, and integration by parts is used to simplify the integrand.

A simple two point Gauss quadrature formula is used to estimate the resulting integrals over each interval.