fem1d_bvp_quadratic
fem1d_bvp_quadratic, a Python code which applies the finite element method, with piecewise quadratic elements, to a two point boundary value problem in one spatial dimension.

The BVP to be solved is:

        -u'' = x * ( x + 3 ) * exp ( x )  over the interval 0 < x < 1
        u(0) = 0.0
        u(1) = 0.0
      
The exact solution is:

        u(x) = x * ( 1 - x ) * exp ( x )
      
To compute a finite element approximation, a set of n equally spaced nodes is defined from 0.0 to 1.0, a set of piecewise quadratic basis functions is set up, with one basis function associated with each node, and then an integral form of the BVP is used, in which the differential equation is multiplied by each basis function, and integration by parts is used to simplify the integrand.

A simple three point Gauss quadrature formula is used to estimate the resulting integrals over each interval.